Changes from dev
Reworked package references to include the entire dependency graph with version ranges. This implements the 'lock file' feature, just not in a separate file (it's in the dproj). This change makes restores reliable, dpm will always restore the versions referenced by the project. It also now support promoting a transient dependency to a direct one.